feat: Detect cursor position in array to sort it
If this works good enough we might add AST parsing for the selection the
same way. That would ensure that even partially selected arrays get properly
detected.
Or we can ditch the selection based detection all together.

Then, since we have a proper AST of the array now, we could (instead of
parsing it using JSON.parse) sort the AST elements and maybe then
transform the  AST back to source code to not change the code
formatting like it is happening now.